refactor : change name of variable to logging
