+++ src/infraestructure/bcrypt/hash.ts
