+++ src/infraestructure/config/LoginDependencie.ts
@@ -1,9 +1,9 @@
 import { AuthenticationService } from "../../application/usecases/authentication/auth-use-case";
 import { UsersRepositoryImpl } from "../db/repository/users/UsersRepositoryImpl";
-import { BcryptHashService } from "../bcrypt/BcryptHashServiceImpl";
-import { loggerService } from "../logger/LoggerService";
+import { BcryptHashService } from "../bcrypt/hash";
+import { logging } from "../logger/logging";
 import { LoginController } from "../../presentation/controllers/authentication/LoginController";
-import { JwtServiceImpl } from "../jwt/JwtServiceImpl";
+import { JwtServiceImpl } from "../jwt/jwt";
 
 const usersRepository = new UsersRepositoryImpl();
 
@@ -11,7 +11,7 @@ const loginService = new AuthenticationService(
   usersRepository,
   new JwtServiceImpl(),
   new BcryptHashService(),
-  loggerService,
+  logging
 );
 
 export const authController = new LoginController(loginService);
