+++ src/infraestructure/config/activity/LIstActivityDependencies.ts
@@ -1,13 +1,13 @@
-import { ListActivityController } from "../../../presentation/controllers/activity/ListActivityController";
+import { ListActivityController } from "../../../presentation/controllers/activity/list-activity-controller";
 import { ListActivityService } from "../../../application/usecases/activity/list-activity-use-case";
 import { ActivityRepositoryImpl } from "../../db/repository/activity/ActivitRepositoryImpl";
-import { loggerService } from "../../logger/LoggerService";
+import { loggerService } from "../../logger/logging";
 
 const activityRepository = new ActivityRepositoryImpl();
 
 const listActivityService = new ListActivityService(
   activityRepository,
-  loggerService,
+  loggerService
 );
 
 const listActivityController = new ListActivityController(listActivityService);
