+++ src/infraestructure/config/activity/UpdateActivityDependencies.ts
@@ -1,16 +1,16 @@
 import { ActivityRepositoryImpl } from "../../db/repository/activity/ActivitRepositoryImpl";
-import { loggerService } from "../../logger/LoggerService";
 import { UpdateActivityService } from "../../../application/usecases/activity/update-activity-use-case";
-import { UpdateActivityController } from "../../../presentation/controllers/activity/UpdateActivityController";
+import { UpdateActivityController } from "../../../presentation/controllers/activity/update-activity-controller";
+import { logging } from "../../logger/logging";
 
 const activityRepository = new ActivityRepositoryImpl();
 
 const updateActivityService = new UpdateActivityService(
   activityRepository,
-  loggerService,
+  logging
 );
 const updateActivityController = new UpdateActivityController(
-  updateActivityService,
+  updateActivityService
 );
 
 export { updateActivityController };
