+++ src/infraestructure/factories/usecases/make-db-add-user.ts
@@ -1,8 +1,8 @@
 import { CreateUserService } from "../../../application/usecases/users/create-user-use-case";
 import { ICreateUserService } from "../../../domain/CreateUserService";
-import { BcryptHashService } from "../../bcrypt/BcryptHashServiceImpl";
+import { BcryptHashService } from "../../bcrypt/hash";
 import { UsersRepositoryImpl } from "../../db/repository/users/UsersRepositoryImpl";
-import { loggerService } from "../../logger/LoggerService";
+import { loggerService } from "../../logger/logging";
 
 export const makeDbAddUser = (): ICreateUserService => {
   const dbAddUser = new UsersRepositoryImpl();
