+++ src/infraestructure/logger/logging.ts
@@ -13,7 +13,7 @@ class LoggerService implements ILogger {
         format.colorize(),
         format.printf(({ level, message, timestamp }) => {
           return `${timestamp} ${level}: ${message}`;
-        }),
+        })
       ),
       transports: [
         new transports.Console(),
@@ -47,4 +47,4 @@ class LoggerService implements ILogger {
   }
 }
 
-export const loggerService = new LoggerService();
+export const logging = new LoggerService();
